Restore integrity of wtl/atldlgs.h
Rounak removed this portion of the header because it was causing build
errors. This was the result of an error fixed in 8463d22.

In my opinion, we should not modify library code unless it's absolutely
necessary. Therefore, I am restoring the original version of this file,
since the modifications are no longer necessary.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>