Build: Fix up project configuration
The true entry point of Explorer is not ExplorerWinMain, but ModuleEntry.
ModuleEntry initialises Explorer's custom "scary crt-wannabe code" and
handles some other actions before and after ExplorerWinMain runs.

FixedBaseAddress was set to false because this is what XP did to avoid
issues with shitty programs. It probably isn't necessary at all, but it is
exactly how XP worked, so I think it's worth using.

Similary, StackCommitSize has been changed to reflect its values in XP.

I'm pretty sure, but not certain, that XP explorer used
WindowsShell.manifest, so I set that as the manifest file.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>