Change ReleaseDebug to use MultiThreadedDLL to fix desktop
There seemed to be strange multithreading issues with the embedded CRT
which can cause the desktop thread to hang for a long time while opening
context menus, and can cause file explorer windows to never open.

It's very strange, but using dynamically-linked CRT seems to fix this so...
CRT bug?

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>